cloud-functions/QR-Redemption-Validator/requirements.txt
google-cloud-datastore
web3~=5.24.0
Flask~=2.0.2
requests~=2.26.0